# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bulbe Energia API
  version: 1.0.0
  description: |
    API REST do sistema Bulbe Energia para monitoramento de energia renovável.

    Funcionalidades:
    - Catálogo de produtos, busca, categorias, ofertas e destaques
    - Carrinho de compras com gerenciamento de itens
    - Cadastro e autenticação de usuários com JWT
    - Pedidos com cálculo de frete, validação de cupons e múltiplos métodos de pagamento
    - Pagamento PIX com geração de QR Code
    - Suporte ao cliente com abertura de chamados
    - Webhook para confirmação de pagamento
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
...
paths:
...
  /api/v1/produtos:
    get:
      tags:
        - Produtos
      summary: Listar ou buscar produtos
      description: Retorna produtos ativos. Quando o parâmetro search é informado, filtra os produtos pelo nome.
      operationId: listarProdutos
      parameters:
        - name: search
          in: query
          required: false
          description: Termo usado para buscar produtos pelo nome.
          schema:
            type: string
          example: lampada
      responses:
        '200':
          description: Lista de produtos retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
...
  /api/v1/produtos/categoria/{categoria}:
    get:
      tags:
        - Produtos
      summary: Listar produtos por categoria
      description: |
        Lista produtos ativos por categoria, com suporte a paginação e ordenação.

        Categorias aceitas: casa, eletronicos, eletrodomesticos e ofertas.
      operationId: listarPorCategoria
      parameters:
        - name: categoria
          in: path
          required: true
          description: Slug da categoria.
          schema:
            type: string
            enum:
              - casa
              - eletronicos
              - eletrodomesticos
              - ofertas
          example: casa
        - name: pagina
          in: query
          required: false
          description: Número da página desejada.
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: limite
          in: query
          required: false
          description: Quantidade de produtos por página.
          schema:
            type: integer
            default: 20
            minimum: 1
          example: 20
        - name: ordenar_por
          in: query
          required: false
          description: Critério de ordenação.
          schema:
            type: string
            enum:
              - preco_asc
              - preco_desc
              - mais_vendidos
              - novidades
          example: preco_asc
      responses:
        '200':
          description: Produtos da categoria retornados com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutosPaginados'
        '400':
          description: Valor inválido para o parâmetro ordenar_por.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroOrdenacao'
        '404':
          description: Categoria não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroCategoria'
  /api/v1/produtos/{id}:
    get:
      tags:
        - Produtos
      summary: Buscar produto por ID
      description: Retorna os dados completos de um produto ativo. Produtos inativos só são retornados quando o header x-admin for true.
      operationId: buscarProdutoPorId
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador do produto.
          schema:
            type: integer
          example: 1
        - name: x-admin
          in: header
          required: false
          description: Permite visualizar produto inativo quando definido como true.
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
          example: 'false'
      responses:
        '200':
          description: Produto encontrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoDetalhado'
        '404':
          description: Produto não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
              example:
                erro: Produto não encontrado
...
  /api/suporte/chamados:
    post:
      tags:
        - Suporte
      summary: Abrir chamado de suporte
      description: Abre um chamado de suporte ao cliente. Envia e-mail de confirmação e notifica equipe interna via Slack.
      operationId: abrirChamado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbrirChamadoRequest'
      responses:
        '201':
          description: Chamado aberto com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChamadoResponse'
        '400':
          description: Campos obrigatórios ausentes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint de login.
  schemas:
...
    Produto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Lâmpada LED 12W
        descricaoLonga:
          type: string
          example: Lâmpada LED econômica com alta durabilidade e baixo consumo de energia.
        galeriaImagens:
          type: array
          items:
            type: string
            format: uri
          example:
            - https://example.com/imagens/lampada1.jpg
        preco:
          type: number
          format: double
          example: 29.9
        desconto:
          type: number
          format: double
          example: 10
        destaque:
          type: boolean
          example: true
        marca:
          type: string
          example: Philips
        categoria:
          type: string
          example: casa
        especificacoesTecnicas:
          type: object
          additionalProperties: true
          example:
            potencia: 12W
            voltagem: 110V
            temperaturaCor: 6500K
        estoque:
          type: integer
          example: 120
        avaliacoesResumidas:
          $ref: '#/components/schemas/AvaliacoesResumidas'
        ativo:
          type: boolean
          example: true
    ProdutoDetalhado:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Lâmpada LED 12W
        descricaoLonga:
          type: string
          example: Lâmpada LED econômica com alta durabilidade e baixo consumo de energia.
        galeriaImagens:
          type: array
          items:
            type: string
            format: uri
        preco:
          type: number
          format: double
          example: 29.9
        desconto:
          type: number
          format: double
          example: 10
        marca:
          type: string
          example: Philips
        categoria:
          type: string
          example: casa
        especificacoesTecnicas:
          type: object
          additionalProperties: true
        estoque:
          type: integer
          example: 120
        avaliacoesResumidas:
          $ref: '#/components/schemas/AvaliacoesResumidas'
    AvaliacoesResumidas:
      type: object
      properties:
        media:
          type: number
          format: double
          example: 4.8
        total:
          type: integer
          example: 256
    ProdutoCategoriaResumo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Lâmpada LED 12W
        preco:
          type: number
          format: double
          example: 29.9
        precoComDesconto:
          type: number
          format: double
          example: 26.91
        imagemPrincipal:
          type: string
          format: uri
          nullable: true
          example: https://example.com/imagens/lampada1.jpg
        categoria:
          type: string
          example: casa
        estoqueDisponivel:
          type: integer
          example: 120
    ProdutosPaginados:
      type: object
      properties:
        categoria:
          type: string
          example: casa
        paginacao:
          $ref: '#/components/schemas/Paginacao'
        produtos:
          type: array
          items:
            $ref: '#/components/schemas/ProdutoCategoriaResumo'
    Paginacao:
      type: object
      properties:
        paginaAtual:
          type: integer
          example: 1
        limite:
          type: integer
          example: 20
        total:
          type: integer
          example: 1
        totalPaginas:
          type: integer
          example: 1

    # ─────────────────────────────────────────
    # CARRINHO
    # ─────────────────────────────────────────
...
    AbrirChamadoRequest:
      type: object
      required:
        - assunto
        - mensagem
        - usuario
      properties:
        assunto:
          type: string
          description: Assunto do chamado.
          example: Problema com meu pedido
        mensagem:
          type: string
          description: Descrição detalhada do problema.
          example: Recebi o produto com defeito e gostaria de solicitar a troca.
        usuario:
          type: object
          required:
            - nome
            - email
          properties:
            nome:
              type: string
              example: João Silva
            email:
              type: string
              format: email
              example: joao@email.com
        pedido_id:
          type: string
          nullable: true
          description: Identificador do pedido relacionado (opcional).
          example: PED-00001
    ChamadoResponse:
      type: object
      properties:
        protocolo:
          type: string
          description: Número de protocolo único do chamado.
          example: SUP-00001-20250520
        status:
          type: string
          enum:
            - aberto
          example: aberto
        assunto:
          type: string
          example: Problema com meu pedido
        criadoEm:
          type: string
          format: date-time
          example: 2025-05-20T10:00:00.000Z
        mensagem:
          type: string
          description: Mensagem de confirmação.
          example: 'Chamado aberto com sucesso. Protocolo: SUP-00001-20250520'

    # ─────────────────────────────────────────
    # WEBHOOK
    # ─────────────────────────────────────────
...
    Erro:
      type: object
      properties:
        erro:
          type: string
          description: Mensagem de erro descritiva.
          example: Mensagem de erro
    ErroCategoria:
      type: object
      properties:
        erro:
          type: string
          example: Categoria 'moveis' não encontrada.
        categoriasDisponiveis:
          type: array
          items:
            type: string
          example:
            - casa
            - eletronicos
            - eletrodomesticos
            - ofertas
    ErroOrdenacao:
      type: object
      properties:
        erro:
          type: string
          example: "Valor de ordenar_por inválido: 'menor_preco'."
        valoresPermitidos:
          type: array
          items:
            type: string
          example:
            - preco_asc
            - preco_desc
            - mais_vendidos
            - novidades